+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # avro-builder changelog
 
+## v0.2.0
+- Add support for `:union` type and DSL method for a record.
+- Make `fixed` syntax more flexible. Both `fixed :f, 7` and `fixed :f, size: 7`
+  are now supported and equivalent.
+
 ## v0.1.0
 - Initial release